+++ eval.py
@@ -1,15 +1,21 @@
 import hydra
+from hydra.core.global_hydra import GlobalHydra
 from omegaconf import OmegaConf
 from torch_points3d.trainer import Trainer
 
 
-@hydra.main(config_path="conf/eval.yaml")
+@hydra.main(config_path="conf", config_name="eval")
 def main(cfg):
     OmegaConf.set_struct(cfg, False)  # This allows getattr and hasattr methods to function correctly
-    print(cfg.pretty())
-
+    if cfg.pretty_print:
+        print(OmegaConf.to_yaml(cfg))
+    
     trainer = Trainer(cfg)
     trainer.eval()
+    #
+    # # https://github.com/facebookresearch/hydra/issues/440
+    GlobalHydra.get_state().clear()
+    return 0
 
 
 if __name__ == "__main__":