+++ torch_points3d/core/data_transform/__init__.py
@@ -53,11 +53,13 @@ def instantiate_transform(transform_option, attr="transform"):
     """
     tr_name = getattr(transform_option, attr, None)
     try:
-        tr_params = transform_option.params
+        # tr_params = transform_option.params
+        tr_params = transform_option.get('params')  # Update to OmegaConf 2.0
     except KeyError:
         tr_params = None
     try:
-        lparams = transform_option.lparams
+        # lparams = transform_option.lparams
+        lparams = transform_option.get('lparams') # Update to OmegaConf 2.0
     except KeyError:
         lparams = None
 
@@ -127,6 +129,41 @@ class LotteryTransform(object):
         return rep
 
 
+class ComposeTransform(object):
+    """
+    Transform to compose other transforms with YAML (Compose of torch_geometric does not work).
+    Example :
+    .. code-block:: yaml
+
+    - transform: ComposeTransform
+      params:
+        transform_options:
+          - transform: GridSampling3D
+            params:
+              size: 0.1
+          - transform: RandomNoise
+            params:
+              sigma: 0.05
+
+
+    Parameters:
+    transform_options: Omegaconf Dict
+        contains a list of transform
+    """
+    def __init__(self, transform_options):
+        self.transform = instantiate_transforms(transform_options)
+
+    def __call__(self, data):
+        return self.transform(data)
+
+    def __repr__(self):
+        rep = "ComposeTransform(["
+        for trans in self.transform.transforms:
+            rep = rep + "{}, ".format(trans.__repr__())
+        rep = rep + "])"
+        return rep
+
+
 class RandomParamTransform(object):
     """
     create a transform with random parameters
@@ -171,6 +208,7 @@ class RandomParamTransform(object):
         the name of the transform
     transform_options: Omegaconf Dict
         contains the name of a variables as a key and min max type as value to specify the range of the parameters and the type of the parameters or it contains the value "value" to specify a variables (see Example above)
+
     """
 
     def __init__(self, transform_name, transform_params):
@@ -184,6 +222,7 @@ class RandomParamTransform(object):
             if "max" in rang and "min" in rang:
                 assert rang["max"] - rang["min"] > 0
                 v = np.random.random() * (rang["max"] - rang["min"]) + rang["min"]
+
                 if rang["type"] == "float":
                     v = float(v)
                 elif rang["type"] == "int":
@@ -196,8 +235,8 @@ class RandomParamTransform(object):
                 dico[p] = v
             else:
                 raise NotImplementedError
-        trans_opt = DictConfig(dict(params=dico, transform=self.transform_name))
 
+        trans_opt = DictConfig(dict(params=dico, transform=self.transform_name))
         random_transform = instantiate_transform(trans_opt, attr="transform")
         return random_transform
 
@@ -207,4 +246,3 @@ class RandomParamTransform(object):
 
     def __repr__(self):
         return "RandomParamTransform({}, params={})".format(self.transform_name, self.transform_params)
-